server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
pydantic-settings==2.7.0
python-multipart==0.0.20
httpx==0.28.1
qdrant-client==1.12.1
openai==1.58.1
azure-identity==1.19.0
beautifulsoup4==4.12.3
requests==2.32.3
python-dotenv==1.0.1
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10
tenacity==9.0.0

# Web Scraping
selenium==4.27.1
webdriver-manager==4.0.2
lxml==5.3.0
readability-lxml==0.8.1
reportlab==4.0.7

# Redis and Celery - Optional for caching and background jobs
# redis==5.2.1
# celery==5.4.0